move config package to pkg directory
