+++ pkg/config/resource.go
