+++ pkg/model/field.go
@@ -19,7 +19,7 @@ import (
 
 	"github.com/gertd/go-pluralize"
 
-	ackgenconfig "github.com/aws-controllers-k8s/code-generator/pkg/generate/config"
+	ackgenconfig "github.com/aws-controllers-k8s/code-generator/pkg/config"
 	"github.com/aws-controllers-k8s/code-generator/pkg/names"
 	"github.com/aws-controllers-k8s/code-generator/pkg/util"
 	awssdkmodel "github.com/aws/aws-sdk-go/private/model/api"
