+++ pkg/model/model.go
@@ -19,8 +19,8 @@ import (
 	"sort"
 	"strings"
 
+	ackgenconfig "github.com/aws-controllers-k8s/code-generator/pkg/config"
 	ackfp "github.com/aws-controllers-k8s/code-generator/pkg/fieldpath"
-	ackgenconfig "github.com/aws-controllers-k8s/code-generator/pkg/generate/config"
 	"github.com/aws-controllers-k8s/code-generator/pkg/generate/templateset"
 	"github.com/aws-controllers-k8s/code-generator/pkg/names"
 	"github.com/aws-controllers-k8s/code-generator/pkg/util"
